requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-cors==3.0.10
python-dotenv==0.19.0
PyPDF2==3.0.1
python-docx==0.8.11
Werkzeug==2.0.1
SQLAlchemy==1.4.23
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
nltk==3.6.3
spacy==3.1.3 